Update BaseLight header and source.

+++ BaseLight.h
@@ -8,13 +8,12 @@ class BaseLight
 {
 public:
 	BaseLight();
-	BaseLight(GLuint shadowWidth, GLuint shadowHeight, glm::vec3 colour, GLfloat ambientIntensity, GLfloat diffuseIntensity);
+	BaseLight(GLfloat shadowWidth, GLfloat shadowHeight, glm::vec3 colour, GLfloat ambientIntensity, GLfloat diffuseIntensity);
 	~BaseLight();
 
 	glm::vec3 GetColour() const { return mAmbientColour; }
 	GLfloat GetAmbientIntensity() const { return mAmbientIntensity; }
 	GLfloat GetDiffuseIntensity() const { return mDiffuseIntensity; }
-
 	ShadowMap* GetShadowMap() const { return mShadowMap; }
 
 protected:
